Fix Service

+++ src/main/java/com/swd392/BatterySwapStation/application/service/StationStaffService.java
@@ -1,6 +1,8 @@
 package com.swd392.BatterySwapStation.application.service;
 
 import com.swd392.BatterySwapStation.domain.entity.StationStaff;
+import com.swd392.BatterySwapStation.domain.enums.UserRole;
+import com.swd392.BatterySwapStation.domain.enums.UserStatus;
 import com.swd392.BatterySwapStation.domain.exception.NotFoundException;
 import com.swd392.BatterySwapStation.domain.repository.StationStaffRepository;
 import lombok.RequiredArgsConstructor;
@@ -14,6 +16,7 @@ import java.util.UUID;
 public class StationStaffService {
 
     private  final StationStaffRepository stationStaffRepository;
+    private final UserService userService;
 
 
     public StationStaff saveStationStaff(StationStaff stationStaff) {
@@ -21,7 +24,16 @@ public class StationStaffService {
     }
 
     public Void deleteStationStaff(StationStaff stationStaff) {
+        var user = userService.getUserById(stationStaff.getStaffId());
+        if (user.getStatus() == UserStatus.ACTIVE){
+            throw new IllegalStateException("Cannot delete active user with ID: " + user.getId());
+        }
+
+        user.setStatus(UserStatus.DELETED);
+        userService.saveUser(user);
+
         stationStaffRepository.delete(stationStaff);
+
         return null;
     }
 
@@ -33,6 +45,7 @@ public class StationStaffService {
     public boolean existsByStaffId(UUID staffId) {
         return stationStaffRepository.existsByStaffId(staffId);
     }
+
     public List<StationStaff> getStaffByStationId(UUID stationId) {
         return stationStaffRepository.findByStation_Id(stationId);
     }
